backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=10.1.0
pydantic>=2.5.0
httpx>=0.25.2
nltk>=3.8.1
textblob>=0.17.1
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=2.0.0
opencv-python>=4.8.0
setuptools>=70.0.0